split up miner-related code into separate modules
much of the miner code is not specific to the OpenGL miner, it shouldn't be in opengl_miner.rs
